old-tests/submission/Infraware/Forms/contents/Forms/button_formaction.html

<!DOCTYPE HTML>
<html>
 <head>
  <title>Forms</title>
  <script src="../../resources/testharness.js"></script>
  <script src="../../resources/testharnessreport.js"></script>
 </head>
 <body>
	<p>
		<h3>button_formAction</h3>
	</p>
	
	<hr>

	<div id="log"></div>

	<form method="post"
		  enctype="application/x-www-form-urlencoded"
		  action=""
		  name="input_form">
	<p><button id='button_id'>button</p>
	</form>

	<script>
		
		var button = document.getElementById("button_id");
		button.formAction = "http://translate.google.com/";

		if (typeof(button.formAction) == "string") {
			test(function() {
				assert_equals(button.formAction, "http://translate.google.com/", "formAction attribute is not correct.");
			});
		} else {
			test(function() {
				assert_unreached("formAction attribute is not exist.");
			});
		}

	</script>

 </body>
</html>
